[Refactor]  #37 - AuthSession관련 코드 분리
